Bootstrapping learnr tutorials image for SciServer
This commit brings in the necessary components to create the learnr tutorial image running on SciServer compute. This is
set to run as the standard idies user, and when the C-MOOR data volume is mounted, tutorial data points there instead of
the included asset data (which are included here as a fallback).